change count type long -> Integer

+++ src/main/java/exastro/Exastro_Days_Tokyo/event_user/repository/vo/CountVO.java
@@ -30,6 +30,6 @@ import lombok.NoArgsConstructor;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class CountVO {
 	
-	private long count;
+	private Integer count;
 	
 }
